Reland 4 commits in ChromeVox darken screen dialog feature.
This reverts commit 0953def82ee77ab44499106c5e87cff462840a30.

Reason for revert: Re-enable for M92+ after disabling for M91.

Original change's description:
> Revert 4 commits in ChromeVox darken screen dialog feature.
>
> Revert "Finalize ChromeVox turn off screen dialog text"
> This reverts commit 3298ae4494e234a768bb7201ab810205502cd01b.
>
> Revert "Update darken screen strings from UX feedback."
> This reverts commit 19b25bbdc61f3196c096b7efbd59ca454e80f686.
>
> Revert "Add 'Toggle darken screen' command to chromevox menu."
> This reverts commit 2174933c47faa45aa41cf9306e978888223b52b6.
>
> Revert "Show a confirmation dialog before first darkening the screen."
> This reverts commit 5241a12f7f2d44bdc342a344714c94f9d208def4.
>
> accelerator command, behavior matches M90.
>
> Bug: 1200484
> AX-Relnotes: Removes darken screen dialog feature in M91.
> Test: Tried ChromeVox darken screen command and another
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2837290
> Commit-Queue: David Tseng <dtseng@chromium.org>
> Cr-Commit-Position: refs/heads/master@{#874392}

Bug: 1200484
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2842028
Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Commit-Queue: Katie Dektar <katie@chromium.org>
Cr-Commit-Position: refs/heads/master@{#874499}

+++ chrome/browser/resources/chromeos/accessibility/chromevox/background/command_handler.js
@@ -244,14 +244,28 @@ CommandHandler.onCommand = function(command) {
     case 'help':
       (new PanelCommand(PanelCommandType.TUTORIAL)).send();
       return false;
-    case 'toggleDarkScreen':
+    case 'toggleScreen':
       const oldState = sessionStorage.getItem('darkScreen');
       const newState = (oldState === 'true') ? false : true;
-      sessionStorage.setItem('darkScreen', (newState) ? 'true' : 'false');
-      chrome.accessibilityPrivate.darkenScreen(newState);
-      new Output()
-          .format((newState) ? '@darken_screen' : '@undarken_screen')
-          .go();
+      if (newState && localStorage['acceptToggleScreen'] !== 'true') {
+        // If this is the first time, show a confirmation dialog.
+        chrome.accessibilityPrivate.showConfirmationDialog(
+            Msgs.getMsg('toggle_screen_title'),
+            Msgs.getMsg('toggle_screen_description'), (confirmed) => {
+              if (confirmed) {
+                sessionStorage.setItem('darkScreen', 'true');
+                localStorage['acceptToggleScreen'] = true;
+                chrome.accessibilityPrivate.darkenScreen(true);
+                new Output().format('@toggle_screen_off').go();
+              }
+            });
+      } else {
+        sessionStorage.setItem('darkScreen', (newState) ? 'true' : 'false');
+        chrome.accessibilityPrivate.darkenScreen(newState);
+        new Output()
+            .format((newState) ? '@toggle_screen_off' : '@toggle_screen_on')
+            .go();
+      }
       return false;
     case 'toggleSpeechOnOrOff':
       const state = ChromeVox.tts.toggleSpeechOnOrOff();
